Moves tree tables into dedicated app
