+++ fig_tree/main/settings.py
@@ -45,6 +45,7 @@ INSTALLED_APPS = [
     'django_bootstrap5',
     'rest_framework',
     'widget_tweaks',
+    'apps.trees',
     'apps.authentication',
     'apps.gen_rest_api',
     'apps.signup',
